pınk/requirements.txt
ccxt==4.1.84
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
python-binance==1.0.19
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.8.2
python-telegram-bot==20.7
